+++ src/Symfony/Component/DependencyInjection/Tests/ChildDefinitionTest.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\DependencyInjection\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ChildDefinition;
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
 
 class ChildDefinitionTest extends TestCase
 {
@@ -130,11 +129,6 @@ class ChildDefinitionTest extends TestCase
         $def->replaceArgument(0, 'foo');
 
         $def->getArgument(1);
-    }
-
-    public function testDefinitionDecoratorAliasExistsForBackwardsCompatibility()
-    {
-        $this->assertInstanceOf(ChildDefinition::class, new DefinitionDecorator('foo'));
     }
 
     /**