+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
  * Removed absolute template paths support in the template name parser
  * Removed support of the `KERNEL_DIR` environment variable with `KernelTestCase::getKernelClass()`.
  * Removed the `KernelTestCase::getPhpUnitXmlDir()` and `KernelTestCase::getPhpUnitCliConfigArgument()` methods.
+ * Removed the "framework.validation.cache" configuration option. Configure the "cache.validator" service under "framework.cache.pools" instead.
 
 3.4.0
 -----